Added modular format to server
